Session 6 Prep

+++ Assets/Common/Scripts/Components/Collectable.cs
@@ -6,5 +6,5 @@ using Unity.Mathematics;
 [GenerateAuthoringComponent]
 public struct Collectable : IComponentData
 {
-    public float points;    
+    public int points;    
 }
